fix(control-ui): hide heartbeat-ack turns from chat history via browser-safe helper (#2770)
Re-port upstream 3f63ba8fd80 (partial): hide assistant HEARTBEAT_OK
acknowledgement turns from the Control-UI chat history on reload, mirroring the
existing silent-reply / empty-user hiding.

The natural helper `isHeartbeatOkResponse` lived behind `heartbeat.ts`, whose
transitive deps (`node:fs`/`node:path` and `../utils.js`) poison the Control-UI
browser bundle — `test-ui-smoke` catches this. Extract the pure token-stripping
logic into a browser-safe `src/auto-reply/heartbeat-token.ts` (no `node:*`
imports, no node-coupled transitive deps); `heartbeat.ts` and `tokens.ts`
re-export it so every server-side caller is unchanged; `heartbeat-filter.ts`
re-points to it and becomes browser-importable. `chat.ts` then hides
heartbeat-ack turns through the same `shouldHideHistoryMessage` path.

Part 1 of #2770 (WhatsApp QR-login sync, upstream 245451b6a9c) is deferred —
the fork lacks the QR-rotation infrastructure the fix hangs off
(`waitForWhatsAppLoginResult`), so making `currentQrDataUrl` non-inert needs the
concurrency-sensitive rotation/render machinery re-architected onto a different
socket-wiring pattern: an all-or-nothing change unsuitable to force for a
LOW-severity, re-scan-recoverable issue. #2770 stays open for it.
