+++ src/auto-reply/heartbeat-token.ts
@@ -1,0 +1,132 @@
+// Browser-safe heartbeat token stripping.
+//
+// This module deliberately has NO node:* imports and no node-coupled transitive
+// deps (it does not import ../utils.js or ./heartbeat.js), so the Control UI
+// bundle can import it — directly or via ./heartbeat-filter.js — without pulling
+// node:fs/node:path into the browser build. `heartbeat.ts` re-exports these
+// symbols so existing server-side callers keep importing them from
+// `./heartbeat.js` unchanged. See #2770.
+
+/** The acknowledgement token emitted by heartbeat-only replies. */
+export const HEARTBEAT_TOKEN = "HEARTBEAT_OK";
+
+export const DEFAULT_HEARTBEAT_ACK_MAX_CHARS = 300;
+
+export type StripHeartbeatMode = "heartbeat" | "message";
+
+// Local copy of escapeRegExp — the shared `../utils.js` home imports node:fs and
+// would poison the browser bundle, so a browser-safe module cannot depend on it.
+function escapeRegExp(value: string): string {
+  return value.replace(/[.*+?^${}()|[\]\\]/g, "\\$&");
+}
+
+function stripTokenAtEdges(raw: string): { text: string; didStrip: boolean } {
+  let text = raw.trim();
+  if (!text) {
+    return { text: "", didStrip: false };
+  }
+
+  const token = HEARTBEAT_TOKEN;
+  const tokenAtEndWithOptionalTrailingPunctuation = new RegExp(
+    `${escapeRegExp(token)}[^\\w]{0,4}$`,
+  );
+  if (!text.includes(token)) {
+    return { text, didStrip: false };
+  }
+
+  let didStrip = false;
+  let changed = true;
+  while (changed) {
+    changed = false;
+    const next = text.trim();
+    if (next.startsWith(token)) {
+      const after = next.slice(token.length).trimStart();
+      text = after;
+      didStrip = true;
+      changed = true;
+      continue;
+    }
+    // Strip the token when it appears at the end of the text.
+    // Also strip up to 4 trailing non-word characters the model may have appended
+    // (e.g. ".", "!!!", "---"). Keep trailing punctuation only when real
+    // sentence text exists before the token.
+    if (tokenAtEndWithOptionalTrailingPunctuation.test(next)) {
+      const idx = next.lastIndexOf(token);
+      const before = next.slice(0, idx).trimEnd();
+      if (!before) {
+        text = "";
+      } else {
+        const after = next.slice(idx + token.length).trimStart();
+        text = `${before}${after}`.trimEnd();
+      }
+      didStrip = true;
+      changed = true;
+    }
+  }
+
+  const collapsed = text.replace(/\s+/g, " ").trim();
+  return { text: collapsed, didStrip };
+}
+
+export function stripHeartbeatToken(
+  raw?: string,
+  opts: { mode?: StripHeartbeatMode; maxAckChars?: number } = {},
+) {
+  if (!raw) {
+    return { shouldSkip: true, text: "", didStrip: false };
+  }
+  const trimmed = raw.trim();
+  if (!trimmed) {
+    return { shouldSkip: true, text: "", didStrip: false };
+  }
+
+  const mode: StripHeartbeatMode = opts.mode ?? "message";
+  const maxAckCharsRaw = opts.maxAckChars;
+  const parsedAckChars =
+    typeof maxAckCharsRaw === "string" ? Number(maxAckCharsRaw) : maxAckCharsRaw;
+  const maxAckChars = Math.max(
+    0,
+    typeof parsedAckChars === "number" && Number.isFinite(parsedAckChars)
+      ? parsedAckChars
+      : DEFAULT_HEARTBEAT_ACK_MAX_CHARS,
+  );
+
+  // Normalize lightweight markup so HEARTBEAT_OK wrapped in HTML/Markdown
+  // (e.g., <b>HEARTBEAT_OK</b> or **HEARTBEAT_OK**) still strips.
+  const stripMarkup = (text: string) =>
+    text
+      // Drop HTML tags.
+      .replace(/<[^>]*>/g, " ")
+      // Decode common nbsp variant.
+      .replace(/&nbsp;/gi, " ")
+      // Remove markdown-ish wrappers at the edges.
+      .replace(/^[*`~_]+/, "")
+      .replace(/[*`~_]+$/, "");
+
+  const trimmedNormalized = stripMarkup(trimmed);
+  const hasToken = trimmed.includes(HEARTBEAT_TOKEN) || trimmedNormalized.includes(HEARTBEAT_TOKEN);
+  if (!hasToken) {
+    return { shouldSkip: false, text: trimmed, didStrip: false };
+  }
+
+  const strippedOriginal = stripTokenAtEdges(trimmed);
+  const strippedNormalized = stripTokenAtEdges(trimmedNormalized);
+  const picked =
+    strippedOriginal.didStrip && strippedOriginal.text ? strippedOriginal : strippedNormalized;
+  if (!picked.didStrip) {
+    return { shouldSkip: false, text: trimmed, didStrip: false };
+  }
+
+  if (!picked.text) {
+    return { shouldSkip: true, text: "", didStrip: true };
+  }
+
+  const rest = picked.text.trim();
+  if (mode === "heartbeat") {
+    if (rest.length <= maxAckChars) {
+      return { shouldSkip: true, text: "", didStrip: true };
+    }
+  }
+
+  return { shouldSkip: false, text: rest, didStrip: true };
+}
